@theckman 2018-01-16
@deepakprabhakara I've not done a deep-dive on the changes here yet, but it looks mostly good from doing a quick skim. Thank you for the contribution.

Would you be able to re-issue this PR against https://github.com/theckman/godspeed so that it can be merged in to my fork?

Two things to note:

* I plan on trying to get @akrylysov's change in first (https://github.com/PagerDuty/godspeed/pull/26), as your work looks to be based off of theirs.
* Before opening a PR against the fork, can you please rebase/squash your commit history against the `master` branch? I only ask because I prefer single (descriptive) commits[1] when introducing changes this this.

[1] https://chris.beams.io/posts/git-commit/