@deepakprabhakara 2018-01-16
@theckman Thank you. Sure thing. I'll submit a PR from your fork squashed into one commit.